# [DNM] storage: use an append-only write batch