Here's an experiment related to #21855, based on a suggestion by @petermattis and @nvanbenschoten.

The C++ MVCCScan implementation is modified to write to a series of buffers with exponentially increasing size, so that there are no copies due to resizing within C++.

C++ then returns this vector of buffers to Go, which then creates a single right-sized buffer and copies all of the C++ memory into the new single buffer.

This seems to make almost no difference in `MVCCScan` benchmarks, but I observed a 40% speedup on a very large COUNT(*). I need to do more benchmarking on this though.



    name                                                   old time/op    new time/op    delta
    MVCCScan_RocksDB/rows=1000/versions=1/valueSize=8-8       369µs ± 1%     375µs ± 1%  +1.64%  (p=0.000 n=10+10)
    MVCCScan_RocksDB/rows=1000/versions=1/valueSize=64-8      436µs ± 3%     437µs ± 5%    ~     (p=0.796 n=10+10)
    MVCCScan_RocksDB/rows=1000/versions=1/valueSize=512-8    2.53ms ±65%    2.76ms ±76%    ~     (p=0.720 n=10+9)

    name                                                   old speed      new speed      delta
    MVCCScan_RocksDB/rows=1000/versions=1/valueSize=8-8    21.7MB/s ± 1%  21.3MB/s ± 1%  -1.62%  (p=0.000 n=10+10)
    MVCCScan_RocksDB/rows=1000/versions=1/valueSize=64-8    147MB/s ± 3%   146MB/s ± 4%    ~     (p=0.796 n=10+10)
    MVCCScan_RocksDB/rows=1000/versions=1/valueSize=512-8  291MB/s ±101%   227MB/s ±83%    ~     (p=0.481 n=10+10)