@petermattis 2018-01-31
@jordanlewis #22244 adds a 10k row case to `BenchmarkMVCCScan`. You might try patching that in to this PR and seeing if it is affected. In my testing, the 1k row benchmarks are not doing enough copying to see a noticeable affect.

Also, the variance in your `valueSize=512` runs is enormous. That isn't typical.