@petermattis 2018-01-31
The approach generally looks good. I'm with holding smaller comments until we work out why this isn't improving the `MVCCScan` benchmarks. Another thing to try is to use the uninitialized byte slice allocation technique from #22244.

---

Review status: 0 of 8 files reviewed at latest revision, 3 unresolved discussions.

---

*[c-deps/libroach/iterator.h, line 36 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/22245#-L4BRbk7ThbwONEm9AIj:-L4BRbk7ThbwONEm9AIk:b-1r7o1t) ([raw file](https://github.com/cockroachdb/cockroach/blob/8035096030b7712445de0eb675dd55112cb3ba54/c-deps/libroach/iterator.h#L36)):*
> ```objc
>       initSlice.len = 16;
>       bufs_.push_back(initSlice);
>       bufPtr_ = initBuf_ + kHeaderSize;
> ```

Seems like it might be easier to return the count of keys separately. I'm guessing you did this for convenience.

---

*[c-deps/libroach/iterator.h, line 38 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/22245#-L4BQj4mVGuLsCWaP1cv:-L4BQj4mVGuLsCWaP1cw:b-dppuhd) ([raw file](https://github.com/cockroachdb/cockroach/blob/8035096030b7712445de0eb675dd55112cb3ba54/c-deps/libroach/iterator.h#L38)):*
> ```objc
>       bufPtr_ = initBuf_ + kHeaderSize;
>     }
> ```

You need to define a destructor here which loops over `bufs_` and frees the memory. 

---

*[c-deps/libroach/iterator.h, line 94 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/22245#-L4BRTnR-AtUL5FNNl3k:-L4BRTnR-AtUL5FNNl3l:b-kbcd05) ([raw file](https://github.com/cockroachdb/cockroach/blob/8035096030b7712445de0eb675dd55112cb3ba54/c-deps/libroach/iterator.h#L94)):*
> ```objc
>   public:
>     char initBuf_[16];
>     int keys_;
> ```

This should either be `int32_t` or `int64_t` depending on what size integer you want.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/22245#-:-L4BR36ZCsFtaXAmdhg8:b-u3rmg6)*
<!-- Sent from Reviewable.io -->