@petermattis 2018-01-31
After fixing the memory leak, I can confirm this provides a significant speedup to `select count(*)` performance. I'm seeing numbers 22% faster than #22244. There are a bunch of changes to this PR that I'd want to see done before merging. @jordanlewis Let's coordinate offline.