@petermattis 2018-02-01
@jordanlewis I can't push to your branch for some reason. Probably user error. Take a look at https://github.com/petermattis/cockroach/tree/pmattis/slicevector. It has 3 additional commits on top of your branch. There is no particular logic to the 3 commits, that was just the order I did things in. I need to give this another pass tomorrow, but everything is looking pretty good at the moment.