@petermattis 2018-02-01
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> :shipit:

---

Review status: 0 of 13 files reviewed at latest revision, 6 unresolved discussions, some commit checks failed.

---

*[c-deps/libroach/chunked_buffer.h, line 51 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/22245#-L4IF5rVP3G7kZttyr9e:-L4IF5rVP3G7kZttyr9f:b-6nv47w) ([raw file](https://github.com/cockroachdb/cockroach/blob/4633a71ef104d82e1c25087367b7d5fd6dcc0d11/c-deps/libroach/chunked_buffer.h#L51)):*
> ```objc
> 
>  private:
>   void put(const char* data, int len, int nextSizeHint);
> ```

`s/nextSizeHint/next_size_hint/g` (we're using lowercase_with_underscores for variable names in C++)

---

*[c-deps/libroach/include/libroach.h, line 258 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/22245#-L4IFhVA346_ybTo2RUO:-L4IFhVA346_ybTo2RUP:bs49fdx) ([raw file](https://github.com/cockroachdb/cockroach/blob/4633a71ef104d82e1c25087367b7d5fd6dcc0d11/c-deps/libroach/include/libroach.h#L258)):*
> ```objc
>   int32_t len;
>   // count is the number of key/value pairs in bufs.
>   int64_t count;
> ```

This should probably be `int32_t` as well.

---

*[c-deps/libroach/include/libroach.h, line 259 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/22245#-L4IFl5BG3SyfaNWo2Pa:-L4IFl5BG3SyfaNWo2Pb:b3mjxpr) ([raw file](https://github.com/cockroachdb/cockroach/blob/4633a71ef104d82e1c25087367b7d5fd6dcc0d11/c-deps/libroach/include/libroach.h#L259)):*
> ```objc
>   // count is the number of key/value pairs in bufs.
>   int64_t count;
> } DBSliceVector;
> ```

Naming nit: perhaps `DBChunkedBuffer`.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/22245#-:-L4IGA-V33CeUW0oDvFv:b-h79mc8)*
<!-- Sent from Reviewable.io -->